Maintain SequenceNumber for the trust bundle of the domain
SPIFFE specs say a `spiffe_sequence` number should be included with the bundle
that gets served by the bundle endpoint. This tries to maintain the sequence
number in the datastore by incrementing it when the bundle gets modified.
Currently on Append when requested and all the time on Prune and only if
we actually made any changes to the bundle.

One thing that I didn't test is how it works with the various UpstreamAuthority
implementations. I think these should be fine based on my understanding
that the UpstreamAuthority just signs certificates and propagates bundle
changes. Multiple servers might end up Appending the same change, but
that should be ok since we only increment when chagnes are made.

With this both the bundle endpoint and `spire-server bundle show
-format=spiffe` show a `spiffe_sequence` field:
```
{
  "keys": [...],
  "spiffe_sequence": 35
}
```

Signed-off-by: Sorin Dumitru <sdumitru@bloomberg.net>