spire-server: Allow specifying entry id when creating entries
Currently spire-server doesn't allow you to specify the entry if when
creating an entry. This makes it harder to manage the entries from
outside spire-server because you need to store the entry id
generated during creation to be able to easily update or delete the
entry.

This doesn't seem like a necessary limitation since you can still
generate globally unique ids even from outside of spire-server.